+++ crates/swc_transformer_import_glob/tests/fixtures/glob_import_negative/input.expected.js
@@ -1,0 +1,3 @@
+const modules = {
+    "./dir/foo.js": ()=>import("./dir/foo.js")
+};