+++ examples/generate-dts/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @farmfe-examples/generate-dts
 
+## 1.0.13
+
+### Patch Changes
+
+- Updated dependencies [032bd4a]
+  - @farmfe/core@0.13.13
+  - @farmfe/js-plugin-dts@0.2.0
+  - @farmfe/cli@0.5.8
+
+## 1.0.12
+
+### Patch Changes
+
+- Updated dependencies [228ca7e]
+  - @farmfe/core@0.13.12
+  - @farmfe/js-plugin-dts@0.2.0
+  - @farmfe/cli@0.5.8
+
+## 1.0.11
+
+### Patch Changes
+
+- Updated dependencies [ac56943]
+  - @farmfe/core@0.13.11
+  - @farmfe/js-plugin-dts@0.2.0
+  - @farmfe/cli@0.5.8
+
 ## 1.0.10
 
 ### Patch Changes