+++ rust-plugins/sass/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @farmfe/plugin-sass
 
+## 0.2.7
+
+### Patch Changes
+
+- 032bd4a: Fix bugs:
+  1. `server.proxy` does not work as expected
+  2. `plugin-css` should treat `xxx.png` as relative path
+  3. `assets` like `/logo.png` under publicDir should be resolved to `publicDir/logo.png`
+
+## 0.2.6
+
+### Patch Changes
+
+- Fix incorrect sass binary resolve for default path
+
 ## 0.2.5
 
 ### Patch Changes